Update RegressionModel.ipynb
Hey Machine Learning Enthusiats! as The Olympics season is arriving soon, here is an exciting Model for you To ease your anxiety and Predict how many medals a country will win based on the previous medals :) We'll be using data from the Olympics, which was originally on Kaggle. You can download the files we'll use in this project here: